Inherited model in website_school_management
