+++ website_school_management/__init__.py
@@ -19,3 +19,4 @@
 #
 ##############################################################################
 from . import controllers
+from . import models
