+++ website_school_management/models/__init__.py
@@ -1,0 +1,1 @@
+from . import program
